Use setenv() instead of putenv()
Using putenv() leads to strange behaviour because it doesn't copy the
buffer passed in, so it is either impossible to correctly clean up the
memory, or liable to get corrupted if you pass an automatic variable.
The correct way is to use setenv().

This also requires defining some preprocessor macros, so we unify those
and set them as global options in the Makefile:

    _DEFAULT_SOURCE
    _SVID_SOURCE
    _XOPEN_SOURCE