Add Wedge
A Wedge is a way to efficiently create a null-terminated C string from
any source, including Slice and Buffer, where that NUL character may be
missing.

If the source contains a "stable" NUL terminator, Wedge just keeps a
Slice to it, otherwise it creates its own copy in a (heap allocated)
Buffer.