Fix UTF-8 decoder -- all praise to tests!
Well, this is embarrassing... Our UTF-8 decoder was just plain wrong: it
did not recognize several invalid encodings.

Because I decided to use the Unicode Replacement Character 0xfffd to
indicate error cases, and decided to expand the tests, I was able to
catch this.  In the end, it was as good a chance as any to switch to a
more robust decoder.  See http://bjoern.hoehrmann.de/utf-8/decoder/dfa/
for details.